+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Download font automatically
 - Add wound counter for GFF
+- Add auto detect json or txt list
+- Removed console input
 
 ## v0.5.0 (2023-04-30)